% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visHiCDB.r
\name{visHiCDB}
\alias{visHiCDB}
\title{Visualization of CDBs or differential CDBs on Hi-C maps}
\usage{
visHiCDB(hicfile, CDBfile, resolution, chr, startloc, endloc, outdir)
}
\arguments{
\item{hicfile}{hicfile is the file names of  the intra-chromosome matrixes.
   The intra-chromosome matrix could be in a dense (a NxN matrix) or sparse (a Kx3 table,Rao et al.) format. 
Show CDBs on one sample,set hicfile as 'SAMPLE_File'.  
Show  differebtial CDBs in two samples, set hicfile as list('SAMPLE1_FILE','SAMPLE2_FILE').}

\item{CDBfile}{CDBfile should be is file name of the CDB files.
     Show CDBs on one sample,set CDBfile as 'SAMPLE_CDB'.  
   Show  differebtial CDBs in two samples, set CDBfile as list('SAMPLE1_CDB','SAMPLE2_CDB').
The CDB file should be formated as the output file of HiCDB.}

\item{resolution}{resolution of Hi-C matrix. This is required.}

\item{chr, startloc, endloc}{numeric observation locus on Hi-C map.This is required.}

\item{outdir}{The output direction. Default will be the directory of  the first sample.}
}
\description{
visHiCDB uses Hi-C raw contact matrix and CDBs as input and outputs figures
  of CDBs or differential CDBs on Hi-C maps
}
\details{
This function outputs a pdf figure showing CDBs on a Hi-C
      map on desired locus. Conserved CDBs are marked as dark blue.
}
\examples{
     ## Show CDB on single Hi-C map
     visHiCDB('sample1/chr18.matrix','CDB1.txt',40000,18,25000000,31150000)
     ## Show differential CDBs
     visHiCDB(list('sample1/chr18.matrix','sample2/chr18.matrix'),
        +list('CDB1.txt','CDB2.txt'),40000,18,25000000,31150000)
}
\author{
Implemented by Fengling Chen

  Any suggestions and remarks might be addressed to Fengling Chen:cfl15@mails.tsinghua.edu.cn
}
